+++ audio/__init__.py
@@ -1,3 +1,7 @@
 from .audio import Audio
+from .exceptions import AudioError
 
-__all__ = ["Audio"]
+__all__ = [
+    "Audio",
+    "AudioError",
+]